+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Add numerical stability to logsumexp function
     - Fix Mean Shift delta hyper-parameter
     - Fix MAPE calculation in Error Analysis report
+    - Fix Swish layer beta parameter gradient
 
 - 2.5.4
     - Fix AdaBoost early stopping algorithm